refactor: extract label.rs
