+++ components/restreamer/src/state.rs
@@ -3,6 +3,7 @@
 
 mod client;
 mod client_statistics;
+mod label;
 mod mixin;
 mod server_info;
 mod settings;
@@ -12,6 +13,7 @@ pub use self::{
     client_statistics::{
         ClientStatistics, ClientStatisticsResponse, StatusStatistics,
     },
+    label::Label,
     mixin::{Delay, Mixin, MixinId, MixinSrcUrl, Volume, VolumeLevel},
     server_info::ServerInfo,
     settings::Settings,
@@ -1833,44 +1835,4 @@ pub enum Status {
 
     /// Failed recently
     Unstable,
-}
-
-/// Label of a [`Restream`] or an [`Output`].
-#[derive(
-    Clone,
-    Debug,
-    Deref,
-    Display,
-    Eq,
-    Hash,
-    Into,
-    PartialEq,
-    Serialize,
-    GraphQLScalar,
-)]
-#[graphql(transparent)]
-pub struct Label(String);
-
-impl Label {
-    /// Creates a new [`Label`] if the given value meets its invariants.
-    #[must_use]
-    pub fn new<'s, S: Into<Cow<'s, str>>>(val: S) -> Option<Self> {
-        static REGEX: Lazy<Regex> =
-            Lazy::new(|| Regex::new(r"^[^,\n\t\r\f\v]{1,70}$").unwrap());
-
-        let val = val.into();
-        (!val.is_empty() && REGEX.is_match(&val))
-            .then(|| Self(val.into_owned()))
-    }
-}
-
-impl<'de> Deserialize<'de> for Label {
-    #[inline]
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        Self::new(<Cow<'_, str>>::deserialize(deserializer)?)
-            .ok_or_else(|| D::Error::custom("Not a valid Label"))
-    }
 }
