+++ components/restreamer/src/state/label.rs
@@ -1,0 +1,46 @@
+use derive_more::{Deref, Display, Into};
+use juniper::GraphQLScalar;
+use once_cell::sync::Lazy;
+use regex::Regex;
+use serde::{de::Error as _, Deserialize, Deserializer, Serialize};
+use std::borrow::Cow;
+
+/// Label of a [`Restream`] or an [`Output`].
+#[derive(
+    Clone,
+    Debug,
+    Deref,
+    Display,
+    Eq,
+    Hash,
+    Into,
+    PartialEq,
+    Serialize,
+    GraphQLScalar,
+)]
+#[graphql(transparent)]
+pub struct Label(String);
+
+impl Label {
+    /// Creates a new [`Label`] if the given value meets its invariants.
+    #[must_use]
+    pub fn new<'s, S: Into<Cow<'s, str>>>(val: S) -> Option<Self> {
+        static REGEX: Lazy<Regex> =
+            Lazy::new(|| Regex::new(r"^[^,\n\t\r\f\v]{1,70}$").unwrap());
+
+        let val = val.into();
+        (!val.is_empty() && REGEX.is_match(&val))
+            .then(|| Self(val.into_owned()))
+    }
+}
+
+impl<'de> Deserialize<'de> for Label {
+    #[inline]
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        Self::new(<Cow<'_, str>>::deserialize(deserializer)?)
+            .ok_or_else(|| D::Error::custom("Not a valid Label"))
+    }
+}
